+++ tests/Fake/ControllerUsingService.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Fake;
+
+final class ControllerUsingService
+{
+    public function __construct(
+        public ServiceWithScalarDependency $service,
+    ) {
+    }
+}